apps/processor/microservices/units/ms-gen-xls/pyproject.toml
[project]
name = "ms-gen-xls"
version = "0.1.0"
description = "Excel Generator microservice – generates Excel reports from templates with placeholder substitution, tables, and charts."
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1",
    "matplotlib>=3.8",
    "grpcio>=1.62",
    "grpcio-tools>=1.62",
    "grpcio-health-checking>=1.62",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "pydantic>=2.0",
    "opentelemetry-api>=1.23",
    "opentelemetry-sdk>=1.23",
    "opentelemetry-instrumentation-grpc>=0.44b0",
    "opentelemetry-exporter-otlp>=1.23",
    "dapr>=1.13",
    "dapr-ext-grpc>=1.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 120
